refactor(chart): 수동 계산식 제거 및 파일 구조 변경
