+++ apps/monitor-web/src/components/charts/index.ts
@@ -1,1 +1,2 @@
 export { default as ApiResponseTimeChart } from "./ApiResponseTimeChart";
+export type { ApiResponseTimeData } from "./types";
