+++ apps/monitor-web/src/components/charts/types.ts
